backend/pyproject.toml
[tool.poetry]
name = "Camp'n'go"
version = "0.1.0"
description = ""
authors = [
    "Patryk Beberok <patryk.beberok0990@gmail.com>",
    "Katarzyna Szczepańczyk <k.szczepanczyk83@gmail.com>",
    "Mateusz Dworaczyk <mateuszdwyk@gmail.com>",
    "Kamil Gajczak <kamil.gajczak@gmail.com>",
]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
django-split-settings = "^1.2.0"
psycopg2 = "2.9.9"
django-cors-headers = "^4.3.1"
argon2-cffi = "^23.1.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.1"
drf-spectacular-sidecar = "^2024.1.1"
django-filter = "^23.5"
gitpython = "^3.1.41"
pillow = "^10.3.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}

[tool.poetry.group.dev.dependencies]
django-silk = "^5.1.0"
bpython = "^0.24"
types-setuptools = "^69.0.0.20240125"
wemake-python-styleguide = "^0.19.2"
flake8 = "^7.1.1"
xenon = "^0.9.3"
bandit = "^1.7.10"
pip-audit = "^2.7.3"
mypy = "^1.13.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
pytest-cov = "^6.0.0"
model-bakery = "^1.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
